requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
importlib-metadata==4.8.2
joblib==1.1.0
lxml==4.6.4
Markdown==3.3.6
mysqlclient==2.1.0
numpy==1.21.4
PyMySQL==1.0.2
python-docx==0.8.11
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
sqlparse==0.4.2
threadpoolctl==3.0.0
zipp==3.6.0
